Use ::class syntax instead of strings in TestFactoryBuilder
This works fine in PHP. The ::class syntax is really only a string,
as before. The class doesn't even need to exist. The main benefit is
that IDEs and many other tools have a chance to understand that these
strings are not just strings, but actually refer to a class name.

Change-Id: Id8b66c382fd5180272b9e94d94270ceb4c69e6c3